Extract common helpers into module
